+++ Editor/Setting/Setting.cs
@@ -1,37 +1,20 @@
 ﻿using System.IO;
 using UnityEditor;
 using UnityEngine;
-using Object = UnityEngine.Object;
 
 namespace AssetLens.Reference
 {
-	internal sealed class Setting : ScriptableObject
+	internal partial class Setting : ScriptableObject
 	{
-		public const uint INDEX_VERSION = 100;
-
-		private static Setting instance = default;
-		
-		private const string k_editorCustomSettingsPath = FileSystem.SettingDirectory + "/AssetLens Setting.asset";
-
 		[SerializeField] private bool enabled = false;
+		
 		[SerializeField] private bool pauseInPlaymode = true;
 		[SerializeField] private bool traceSceneObject = false;
 		[SerializeField] private bool useEditorUtilityWhenSearchDependencies = false;
 		[SerializeField] private bool displayIndexerVersion = false;
 		[SerializeField] private string localization = "English";
 		
-		internal static Setting Inst {
-			get => GetOrCreateSettings();
-		}
-
-		public static bool IsEnabled {
-			get => GetOrCreateSettings().enabled;
-			set
-			{
-				GetOrCreateSettings().enabled = value;
-				EditorUtility.SetDirty(GetOrCreateSettings());
-			}
-		}
+		[SerializeField] private bool useUIElementsWindow = false;
 		
 		public static bool PauseInPlaymode => GetOrCreateSettings().pauseInPlaymode;
 		
@@ -62,57 +45,8 @@ namespace AssetLens.Reference
 			}
 		}
 
-		private static Setting GetOrCreateSettings()
-		{
-			if (instance != null)
-			{
-				return instance;
-			}
-
-			instance = EditorGUIUtility.Load(FileSystem.SettingDirectory + "/ReferenceSetting.asset") as Setting;
-			if (instance != null)
-			{
-				AssetDatabase.MoveAsset(AssetDatabase.GetAssetPath(instance), k_editorCustomSettingsPath);
-				AssetDatabase.SaveAssets();
-				
-				return instance;
-			}
-
-			instance = EditorGUIUtility.Load(k_editorCustomSettingsPath) as Setting;
-
-			if (instance == null)
-			{
-				instance = CreateInstance<Setting>();
-				instance.enabled = false;
-
-				if (!Directory.Exists(FileSystem.SettingDirectory))
-				{
-					Directory.CreateDirectory(FileSystem.SettingDirectory);
-					AssetDatabase.ImportAsset(FileSystem.SettingDirectory);
-				}
-
-				AssetDatabase.CreateAsset(instance, k_editorCustomSettingsPath);
-				AssetDatabase.SaveAssets();
-			}
-
-			return instance;
-		}
-
-		internal class AssetDataSettingsProviderRegister
-		{
-			[SettingsProvider]
-			public static SettingsProvider CreateFromSettingsObject()
-			{
-				Object settingsObj = GetOrCreateSettings();
-				
-				AssetSettingsProvider provider =
-					AssetSettingsProvider.CreateProviderFromObject($"Project/Asset Lens", settingsObj);
-
-				provider.keywords =
-					SettingsProvider.GetSearchKeywordsFromSerializedObject(
-						new SerializedObject(settingsObj));
-				return provider;
-			}
+		public static bool UseUIElements {
+			get => GetOrCreateSettings().useUIElementsWindow;
 		}
 	}
 }